adding entry_point mechanism
