+++ src/genie/libs/parser/utils/__init__.py
@@ -1,1 +1,3 @@
 from .common import get_parser, get_parser_exclude, get_parser_commands
+from . import entry_points
+
